Unifying Wikimedia list generator for DEK .svg
DEK_wikimedia_preformat.py's output, [wikimedia_addresses.txt], is
the sole -- yet mandatory -- parameter to generate all three types
of lists previously requested to ease the construction of the DEK
dictionary by this one script.  So it is simpler in use, because
now only one instead of three scripts have to be called, and to be
maintained to create the assistance.

In addition, by same token as introduced to yield the input file,
[wikimedia_addresses.txt], the three outputs this new script writes
equally have a header to ease the files' management with git.