Preformatter yields word-list ordered addresses
The previous form of DEK_wikimedia_preformat.py simply provided a
list of addresses for the .svg on Wikimedia's servers.  To trace
changes, however, it may be helpful if these addresses are sorted
by the same order of consecution as by DEK_wikimedia_lists.py in
file wikimedia_wordlist.txt.