deploy argparse to use DEK_wikimedia_lists.py
The previous version of the script required two interactions by
the user: launch of the script, and definition of the type of file
to write.  Simultaneous launch of the script and provision of the
parameter to generate either a wordlist, the list of previews, or
the grouped list of previews seems to be a better approach in the
long run.